季度选择器改为基于DateRangePicker配置
